requirements.txt
langchain
langchain_openai
langchain_community
PyPDF2
streamlit
python-dotenv
neo4j
openai  # Se necessário para integrar com a API OpenAI
